+++ Libraries/PeriphDrivers/Source/TMR/tmr_revc.h
@@ -56,6 +56,6 @@ void MXC_TMR_RevC_ClearFlags(mxc_tmr_regs_t *tmr);
 uint32_t MXC_TMR_RevC_GetFlags(mxc_tmr_regs_t *tmr);
 void MXC_TMR_RevC_SetCompare(mxc_tmr_regs_t *tmr, uint32_t cmp_cnt);
 void MXC_TMR_RevC_SetCount(mxc_tmr_regs_t *tmr, uint32_t cnt);
-void MXC_TMR_RevC_TO_Start(mxc_tmr_regs_t *tmr, unsigned long us);
+void MXC_TMR_RevC_TO_Start(mxc_tmr_regs_t *tmr, uint32_t us);
 int MXC_TMR_RevC_GetTime(mxc_tmr_regs_t *tmr, uint32_t ticks, uint32_t *time,
                          mxc_tmr_unit_t *units);
